feat/#78: favicon, metadata 설정
