+++ app/(main)/layout.tsx
@@ -1,13 +1,7 @@
 import pretendard from '../utils/font';
-import type { Metadata } from 'next';
 import './../globals.css';
 import SideBar from '../components/SideBar';
 import TopBar from '../components/TopBar';
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-};
 
 export default function RootLayout({
   children,
