+++ app/layout.tsx
@@ -1,10 +1,13 @@
-import pretendard from "./utils/font";
-import type { Metadata } from "next";
-import "./globals.css";
+import pretendard from './utils/font';
+import type { Metadata } from 'next';
+import './globals.css';
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: 'Xpact',
+  description: '작은 경험도 강력한 임팩트로, Xpact',
+  icons: {
+    icon: '/images/favicon.png',
+  },
 };
 
 export default function RootLayout({
